Relax metadata format checks to ignore Provides.
With over two thousand builds in our environment we've found a lot of
packages that incorrectly declare "Provides" in their metadata.
We felt that a pragmatic trade-off of ignoring Provides with 2.1
metadata format (as was already done with Requires) is the best approach
that allowed us to just enforce a single package hotfix -- distlib,
rather than making hotfixes of dozens of old releases.

In addition, two tests were made a little more robust.
One of them assumed that test runner's umask is 022, which is not
necessarily always the case. The corrected test takes care to preserve
the umask while setting the default and restore it at the end of the
test. The other test needed correction for realpath on macOS with Python
3.