scriptmaker: fix regression that prevented proper shebang when building python
In commit ad0ea221431689cb82c89b4942228fb07a87784d a special case was
made to avoid the versioned python executable shebang in cases of
sysconfig.is_python_build(), if and only if the resulting executable
does not exist on disk. The stated rationale was for python builds from
source on Windows, but the commit message claims it is ARM64-specific.
The latter is untrue, and I cannot verify the former -- but it is
certainly broken on platforms *other* than Windows, because when
is_python_build() is true, BINDIR is where python would eventually
install itself to, and the only thing guaranteed to exist is the
in-progress build centered on sys.executable.

So the check always failed, and we never ended up using a versioned
executable.

Assuming we need to handle Windows by not using a versioned executable,
the appropriate way to handle this is to actually check for Windows,
rather than to check for the executable and assume this tells us if we
are on Windows.

Fixes https://github.com/python/cpython/issues/98183